validate uploaded file when updating a media entry also fixed typo on language file

+++ src/Platform/Media/Repositories/DbMediaRepository.php
@@ -210,28 +210,35 @@ class DbMediaRepository implements MediaRepositoryInterface {
 
 		if ($file instanceof UploadedFile)
 		{
-			// Delete the old media file
-			Media::delete($model->path);
+			if ($this->validForUpload($file))
+			{
+				// Delete the old media file
+				Media::delete($model->path);
 
-			File::delete(media_cache_path($model->thumbnail));
+				File::delete(media_cache_path($model->thumbnail));
 
-			// Upload the new file
-			$uploaded = Media::upload($file);
+				// Upload the new file
+				$uploaded = Media::upload($file);
 
-			Event::fire('platform.media.uploaded', array($model, $uploaded, $file));
+				Event::fire('platform.media.uploaded', array($model, $uploaded, $file));
 
-			$imageSize = $uploaded->getImageSize();
+				$imageSize = $uploaded->getImageSize();
 
-			// Update the media entry
-			$model->fill(array(
-				'path'      => $uploaded->getPath(),
-				'extension' => $uploaded->getExtension(),
-				'mime'      => $uploaded->getMimetype(),
-				'size'      => $uploaded->getSize(),
-				'is_image'  => $uploaded->isImage(),
-				'width'     => $imageSize['width'],
-				'height'    => $imageSize['height'],
-			));
+				// Update the media entry
+				$model->fill(array(
+					'path'      => $uploaded->getPath(),
+					'extension' => $uploaded->getExtension(),
+					'mime'      => $uploaded->getMimetype(),
+					'size'      => $uploaded->getSize(),
+					'is_image'  => $uploaded->isImage(),
+					'width'     => $imageSize['width'],
+					'height'    => $imageSize['height'],
+				));
+			}
+			else
+			{
+				return false;
+			}
 		}
 
 		$model->fill($data);
